+++ CHANGES.md
@@ -6,6 +6,7 @@
 - Added new setting "useEscapeKeyToExitFullScreenMode", which states if the Escape key should exit full-screen mode (if enabled).
 - Added global Ctrl/Command + Shift + F11 support to toggle between full screen mode (if enabled).
 - Fixed a fault that caused the display to refresh and tooltips to be shown when one of the drop-down menus is open.
+- In the Full Day view, when clicking an event that overlaps another, it will now be brought to the front.
 - 
 
 ### Version 1.4.3: